--- app/app.js
@@ -1,51 +1,0 @@
-import { useState } from "react";
-import { View, ScrollView, SafeAreaView, Text } from "react-native";
-import { Stack, useRouter } from "expo-router";
-
-import { COLORS, icons, images, SIZES } from "../constants";
-import {
-  Nearbyjobs,
-  Popularjobs,
-  ScreenHeaderBtn,
-  Welcome,
-} from "../components";
-
-const Home = () => {
-  const router = useRouter();
-
-  return (
-    // Body general
-      <SafeAreaView style={{ flex: 1, backgroundColor: COLORS.lightWhite }}>
-        <Stack.Screen
-          options={{
-            // NavBar
-            headerStyle: { backgroundColor: COLORS.gray },
-            headerShadowVisible: false,
-            headerLeft: () => (
-              <ScreenHeaderBtn iconUrl={icons.menu} dimension="60%" />
-            ),
-            headerRight: () => (
-              <ScreenHeaderBtn iconUrl={images.profile} dimension="100%" />
-            ),
-            headerTitle: "                     Ligma",
-          }}
-        />
-
-        <ScrollView showsVerticalScrollIndicator={false}>
-          <View
-            style={{
-              flex: 1,
-              padding: SIZES.small,
-            }}
-          >
-            <Welcome />
-            <Popularjobs />
-            <Nearbyjobs />
-
-          </View>
-        </ScrollView>
-      </SafeAreaView>
-  );
-};
-
-export default Home;
